Accept an options Hash in #playlist
Previous implementation only accepted playlist_id and order parameter.
This prevented fetching all videos in a playlist. I changed #playlist
implementation to accept playlist_id and an opts Hash.

Now 'start-index' and 'max-results' can be passed in the hash.

Note: this change is not backwards-compatible with code using the older
order parameter.